\name{taxa}
\alias{taxa}
\alias{reference}
\docType{data}
\title{
Example list of taxon names
}
\description{
Object \code{taxa} contains taxon names for 50 alpine species determined to the lowest possible taxonomic level. Taxonomy of vascular plants, bryophytes and lichens according to Fischer et al. (2008), Köckinger et al (2011) and Türk & Hafellner (2010).
Object \code{reference} contains a subset of species selected from an exemplary taxonomic reference list (Hennekens & Joop, 2009).
}
\usage{data(taxa)}
\format{
  Two objects, each a vector of characters.
}
\references{
Türk, R. \& Hafellner, J. (2010). Nachtrag zur Bibliographie der Flechten in Österreich. In Ehrendorfer, F., editor, Biosystematics and Ecology Series, volume 27. Austrian Academy of Sciences Press.

Fischer, M., Oswald, K., & Adler, W. (2008). Exkursionsflora fÜr Österreich, Liechtenstein und Südtirol, Volume 3. Biologiezentrum der Oberösterreichischen Lansdesmuseen.

Köckinger, H., Schröck, C., Krisai, R., \& Zechmeister, H. G. (2011) Checklist of austrian bryophytes. See the \href{http://131.130.59.133/projekte/moose/}{webpage}.

Hennekens, S. M. & Joop, H. J. (2009). Turboveg, a comprehensive data base management system for vegetation data. Journal of Vegetation Science, 12:589–591.

}
\examples{
data(taxa)
str(taxa)
str(reference)
}
\keyword{ datasets }
