\name{ the plant list}
\alias{tplFamily}

\title{ Taxonomix searches }

\description{
Search for the family of a given species.
}

\usage{
tplFamily(x)
}

\arguments{
  \item{ x }{
a character.
}
}

\details{
Only the genus is queried from the plant list server.
}
\value{
a character.
}

\author{
Roland Kaiser
}
\references{
Kalwij, J. M. (2012). Review of ‘the plant list, a working list of all plant species’. Journal of Vegetation Science, 23:998–1002.
}
\seealso{
\code{\link[Taxonstand:TPL]{Taxonstand}}
}
\examples{
# Nothing here yet
}
\keyword{ manip }
