\name{prune}
\alias{pruneGenus}

\title{ Prune taxa }

\description{
Prune Genus from taxon string
}

\usage{
pruneGenus(x)
}

\arguments{
  \item{ x }{
a character.
}
}

\details{
Simple minded utility function.
}
\value{
a character.
}

\author{
Roland Kaiser
}
\seealso{
\code{\link{isWhat}}
}
\examples{
# nOthing here yet
}
\keyword{ manip }
